refactor(runtime): extract shared web and tauri runtime helpers
