+++ apps/gpt-image-2-app/src-tauri/Cargo.toml
@@ -19,6 +19,7 @@ tauri-build = { version = "2.0.0", features = [] }
 
 [dependencies]
 gpt-image-2-core = { path = "../../../crates/gpt-image-2-core", version = "0.7.1" }
+gpt-image-2-runtime = { path = "../../../crates/gpt-image-2-runtime", version = "0.7.1" }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
 tauri = { version = "2.0.0", features = ["protocol-asset", "image-png"] }
