+++ apps/gpt-image-2-app/src-tauri/src/lib.rs
@@ -1,5 +1,5 @@
 use std::{
-    collections::{BTreeMap, VecDeque},
+    collections::BTreeMap,
     fs,
     path::{Path, PathBuf},
     process::Command,
