+++ crates/gpt-image-2-runtime/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "gpt-image-2-runtime"
+version = "0.7.1"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+authors.workspace = true
+description = "Shared runtime helpers for GPT Image 2 web and Tauri queues."
+
+[package.metadata.dist]
+dist = false
+
+[dependencies]
+gpt-image-2-core = { path = "../gpt-image-2-core", version = "0.7.1" }
+serde_json = "1.0.145"
