+++ crates/gpt-image-2-web/src/main.rs
@@ -1,10 +1,10 @@
 use std::{
-    collections::{BTreeMap, VecDeque},
+    collections::BTreeMap,
     env, fs,
     path::{Path as FsPath, PathBuf},
     sync::{Arc, Mutex, mpsc},
     thread,
-    time::{SystemTime, UNIX_EPOCH},
+    time::SystemTime,
 };
 
 use axum::{
